refactor: Update API hooks to consolidate query and mutation imports under a single 'queries' and 'mutations' directory structure for improved organization and maintainability
