+++ src/components/molecules/PostCard/PostCardDropdown.tsx
@@ -5,7 +5,7 @@ import Icon from '@/components/atoms/Icon';
 import IconButton from '@/components/atoms/IconButton';
 import { ROUTES } from '@/constants/routes';
 import type { Post } from '@/db/schema';
-import { useDeletePostMutation } from '@/hooks/api/posts/useDeletePostMutation';
+import { useDeletePostMutation } from '@/hooks/api/posts/mutations';
 import { useRouter } from '@/i18n/routing';
 import dynamic from 'next/dynamic';
 import { useTranslations } from 'next-intl';
