+++ src/components/organisms/AuthModal/AuthModalSignUpForm/useAuthModalSignUpForm.tsx
@@ -1,5 +1,4 @@
-import { useSignInMutation } from '@/hooks/api/auth/useSignInMutation';
-import { useSignUpMutation } from '@/hooks/api/auth/useSignUpMutation';
+import { useSignInMutation, useSignUpMutation } from '@/hooks/api/auth/mutations';
 import { useAppForm } from '@/hooks/useFormHook';
 import type { SignUpValues } from '@/schema/auth';
 import { signUpSchema } from '@/schema/auth';
