+++ src/hooks/api/posts/queries/useGetPostQuery.ts
@@ -28,6 +28,8 @@ export const useGetPostQuery = ({ id, enabled = true }: Props) => {
         return allPosts.find((post) => post.id === id);
       }
     },
+    staleTime: 30 * 1000, // 30 seconds
+    gcTime: 0,
   });
 
   return {
