+++ src/hooks/api/profile/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './useGetUserByScreenNameQuery';
